Add Win32 support for timing functions
The C standard library only provides second-resolution wall clock timing
functions. On Linux the POSIX timing APIs are available to provide
millisecond-resolution timing but these APIs are not available on
Windows. Windows provides similar functions in the Win32 API.

This commit adds a preprocessor check that causes the POSIX API to be
used for timing when compiling for Linux and the Win32 API to be used
when compiling for Windows.